Added refactoring
